+++ semiProject/.classpath
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
 	<classpathentry kind="src" path="src/main/java"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-11">
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-17">
 		<attributes>
 			<attribute name="module" value="true"/>
 		</attributes>
@@ -13,15 +13,15 @@
 	</classpathentry>
 	<classpathentry kind="con" path="org.eclipse.jst.j2ee.internal.web.container"/>
 	<classpathentry kind="con" path="org.eclipse.jst.j2ee.internal.module.container"/>
-	<classpathentry kind="lib" path="/Users/ijin/Documents/tools/lib/activation-1.1.jar"/>
-	<classpathentry kind="lib" path="/Users/ijin/Documents/tools/lib/cos.jar"/>
-	<classpathentry kind="lib" path="/Users/ijin/Documents/tools/lib/gson-2.9.0.jar"/>
-	<classpathentry kind="lib" path="/Users/ijin/Documents/tools/lib/javax.mail-1.6.2.jar"/>
-	<classpathentry kind="lib" path="/Users/ijin/Documents/tools/lib/json-simple-1.1.1.jar"/>
-	<classpathentry kind="lib" path="/Users/ijin/Documents/tools/lib/lombok.jar"/>
-	<classpathentry kind="lib" path="/Users/ijin/Documents/tools/lib/ojdbc11-21.5.0.0.jar"/>
-	<classpathentry kind="lib" path="/Users/ijin/Documents/tools/lib/taglibs-standard-impl-1.2.5.jar"/>
-	<classpathentry kind="lib" path="/Users/ijin/Documents/tools/lib/taglibs-standard-jstlel-1.2.5.jar"/>
-	<classpathentry kind="lib" path="/Users/ijin/Documents/tools/lib/taglibs-standard-spec-1.2.5.jar"/>
+	<classpathentry kind="lib" path="C:/tools/lib/activation-1.1.jar"/>
+	<classpathentry kind="lib" path="C:/tools/lib/cos.jar"/>
+	<classpathentry kind="lib" path="C:/tools/lib/gson-2.9.0.jar"/>
+	<classpathentry kind="lib" path="C:/tools/lib/javax.mail-1.6.2.jar"/>
+	<classpathentry kind="lib" path="C:/tools/lib/json-simple-1.1.1.jar"/>
+	<classpathentry kind="lib" path="C:/tools/lib/lombok.jar"/>
+	<classpathentry kind="lib" path="C:/tools/lib/ojdbc11-21.5.0.0.jar"/>
+	<classpathentry kind="lib" path="C:/tools/lib/taglibs-standard-impl-1.2.5.jar"/>
+	<classpathentry kind="lib" path="C:/tools/lib/taglibs-standard-jstlel-1.2.5.jar"/>
+	<classpathentry kind="lib" path="C:/tools/lib/taglibs-standard-spec-1.2.5.jar"/>
 	<classpathentry kind="output" path="src/main/webapp/WEB-INF/classes"/>
 </classpath>
