+++ semiProject/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+encoding//src/main/webapp/WEB-INF/member/login.jsp=UTF-8
