+++ semiProject/src/main/java/edu/kh/community/member/controller/MyPageEnter.java
@@ -26,9 +26,8 @@ public class MyPageEnter extends HttpServlet {
 	
 	 protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		 HttpSession session = request.getSession();
-			
-			
-			
+	
+			System.out.println(session);
 			request.getRequestDispatcher("/WEB-INF/member/mypage.jsp").forward(request, response);
 }
 }
