+++ semiProject/src/main/java/edu/kh/community/sql/member-sql.xml
@@ -47,9 +47,9 @@ AND MEM_ESCAPE = 'N'
 	<!-- 회원 탈퇴 -->
 	<entry key="secession">
 		UPDATE MEMBER SET
-		SECESSION_FL = 'Y'
-		WHERE MEMBER_NO = ?
-		AND MEMBER_PW = ?
+		MEM_ESCAPE = 'Y'
+		WHERE MEM_NO = ?
+		AND MEM_PW = ?
 	</entry>
