+++ semiProject/src/main/webapp/WEB-INF/member/login.jsp
@@ -10,7 +10,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name ="google-signin-client_id" 
     content="935814438564-8ocf3d4e2vsp790u21l0vctt9qahq9bj.apps.googleusercontent.com">
-    <link rel="stylesheet" href="../resources/css/로그인.css">
+    <link rel="stylesheet" href="../resources/css/login.css">
     <script src="https://code.jquery.com/jquery-3.6.0.js" 
     integrity="sha256-H+K7U5CnXl1h5ywQfKtSj8PCmoN9aaq30gDh27Xc0jk=" 
     crossorigin="anonymous"></script>
@@ -105,7 +105,7 @@
 
     </div>
 
-    <script src="../resources/js/로그인.js"></script>
+    <script src="../resources/js/login.js"></script>
     
     <script src="https://developers.kakao.com/sdk/js/kakao.js"></script>
     <script>
