--- semiProject/src/main/webapp/WEB-INF/member/회원가입.jsp
@@ -1,74 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-    
-    
-    <!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/css/회원가입(1).css">
-    
-    <script src="https://kit.fontawesome.com/47910b9f42.js" crossorigin="anonymous"></script>
-    
-    <title>회원가입(1)</title>
-</head>
-<body>
-    <div id="movieSourcing">
-        <div id="layout">
-            <div id="root">
-                <header>
-                    <div>
-                        <a href="#" class="backpage">
-                            <i class="fa-solid fa-chevron-left" id="backpageimg"></i>
-                        </a>
-                    </div>
-                </header>
-                <main id="contents">
-                    <section class="signup">
-                        <span class="top">회원 가입</span>
-
-                    </section>
-                    <section class="infomationbox">
-                        <div class="info">
-                            <div class="infobox">
-                                <span class="info1">약관동의</span>
-                                <div>
-                                &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;<span class="info2">가입을 위해 다음 정책에 대한 동의가 필요합니다.</span>
-                                </div>
-                                
-                                <div>
-                                    <input type="checkbox" id="checkAll"> <span class="info3">모두 동의합니다.</span>
-                                </div>
-                                <hr>
-                                <div>
-                                    <input type="checkbox" id="check1"> <span class="info3">[필수] 이용약관에 동의합니다.</span>
-                                </div>
-                         
-                                <div>
-                                    <input type="checkbox" id="check2"> <span class="info3">[필수] 개인정보 수집 및 이용에 동의합니다.</span>
-                                </div>
-                         
-                                <div>
-                                    <input type="checkbox" id="check3"> <span class="info3">[필수] 본인은 만 14세 이상입니다.</span>
-                                </div>
-                          
-                                <div>
-                                    <input type="checkbox" id="check4"> <span class="info3">[선택] 시사회 및 이벤트 등 알림 수신에 동의합니다.</span>
-                                </div>
-                            </div>
-                        </div>
-                    </section>
-                   
-                    <section class="nextbox">
-                        <button>다음</button>
-                    </section>
-                </main>
-            </div>
-        </div>
-
-    </div>
-    
-</body>
-</html>
