+++ semiProject/src/main/webapp/resources/css/login.css
@@ -1,0 +1,251 @@
+*{
+    margin : 0;
+    padding : 0;
+    box-sizing: border-box;
+}
+
+a{
+    text-decoration: none;
+    color: white;
+}
+span {
+    color: white;
+}
+
+#root {
+    position : relative;
+    min-width : 320px;
+    background-color: black;
+}
+
+#contents {
+    position : relative;
+    min-width : 500px;
+    max-width: 600px;
+
+    background-color: black;
+    
+    margin : 0 auto;
+    box-sizing: border-box;
+    overflow : hidden;
+}
+
+header {
+    position : relative;
+    min-width : 320px;
+    background-color: black;
+    display : flex;
+    align-items: center;
+}
+header > div {
+    position : relative;
+    
+    max-width : 800px;
+    display : flex;
+}
+.backpage {
+    margin-top: 30px;
+    margin-left: 30px;
+    display : flex;
+}
+#backpageimg {
+    font-size: 30px;
+}
+
+section {
+    display : block;
+}
+
+.login > div > span {
+    font-weight: bold;
+    font-size: 16px;
+    color: rgb(210, 210, 210);
+}
+
+.logintop {
+    display : flex;
+    text-align: center;
+    margin : 16px auto 30px;
+    justify-content: center;
+}
+
+.top {
+    font-size: 24px;
+    font-weight: bold;
+}
+
+fieldset {
+    border-style: none;
+
+}
+
+.login {
+    padding : 30px 50px;
+    margin : 8px 0;
+    border-radius: 15px;
+    background-color:  rgb(57, 41, 41);
+
+}
+.loginwrap {
+    display : flex;
+    flex-direction: column;
+    margin : 25px 0 0;
+}
+
+
+.id{
+    margin : 14px 0; 
+    display : flex;
+    justify-content: left;
+}
+
+.id > input {
+    width : 85%;
+    height : 50px;
+
+    margin: auto;
+    font-size: 20px;
+    font-weight: bold;
+    border-radius: 5px;
+}
+
+
+.password{
+    margin : 14px 0; 
+    display : flex;
+    justify-content: left;
+}
+
+.password > input {
+    width : 85%;
+    height : 50px;
+
+    margin: auto;
+    font-size: 20px;
+    font-weight: bold;
+    border-radius: 5px;
+}
+
+
+input::placeholder {
+    font-size: 16px;
+    padding-left: 15px;
+}
+.btnwrap {
+    display : flex;
+    flex-direction: column;
+
+}
+
+.btnwrap > div {
+    margin : 14px auto;
+    border-radius: 30px;
+    height: 55px;
+    width: 75%;
+}
+
+#nomalLogin {
+    display: flex;
+    text-align: center;
+    justify-content: center;
+    line-height: 55px;
+    background-color: rgb(181, 133, 128);
+    cursor: pointer;
+
+    margin : 14px auto;
+    border-radius: 30px;
+    height: 55px;
+    width: 75%;
+}
+#nomalLogin > span {
+    font-size: 20px;
+    font-weight: bold;
+}
+
+#kakaoLogin {
+    display: flex;
+    text-align: center;
+    justify-content: space-between;
+    line-height: 55px;
+    background-color: rgb(247, 205, 47);
+    cursor: pointer;
+}
+#kakaoLogin > div > span {
+    font-size: 20px;
+    color: rgb(57, 41, 41);
+    font-weight: bold;
+}
+#kakaoLogin > div:nth-of-type(1){
+    width: 15%;
+    margin: auto 0;
+    line-height: 42px;
+
+}
+#kakaoLogin > div:nth-of-type(2){
+    width: 70%;
+}
+#kakaoLogin > div:nth-of-type(3){
+    width: 15%;
+}
+#kakaoImg {
+    color: black;
+    margin: auto;
+    font-size: 40px;
+    line-height: 55px;
+    margin-left: 23px;
+    color :rgb(57, 41, 41);
+}
+
+#googleLogin {
+    display: flex;
+    text-align: center;
+    justify-content: space-between;
+    line-height: 55px;
+    background-color: black ;
+    cursor: pointer;
+    
+}
+
+#googleLogin > div:nth-of-type(1){
+    width: 15%;
+    margin: auto 0;
+    line-height: 42px;
+
+}
+#googleLogin > div:nth-of-type(2){
+    width: 70%;
+}
+#googleLogin > div:nth-of-type(3){
+    width: 15%;
+}
+
+#googleLogin > div > span {
+    font-size: 20px;
+    font-weight: bold;
+}
+
+.googleBox {
+    height: 44px;
+    width: 44px;
+    border-radius: 50%;
+    background-color: white;
+    margin: auto 0;
+    line-height: 42px;
+    margin-left: 23px;
+}
+
+#googleImg {
+    color: black;
+    margin: auto;
+    font-size: 34px;
+    line-height: 44px;
+}
+.enderMenu {
+    margin : 0;
+    display: flex;
+    justify-content: space-between;
+}
+
+.under {
+    padding: 40px;
+}
